refactor: moved exceptionhandler and exceptions to own package
